# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SISTEMA EDUPOLI COMPONENTE BACKEND",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://proyectoedupoli.herokuapp.com/"
        }
    ],
    "paths": {
...
        "/api/v1/carreras/admin": {
            "get": {
                "tags": [
                    "Gestion Carreras"
                ],
                "summary": "Lista de carreras activas e inactivas",
                "description": "Carreras a visualizar",
                "operationId": "Carreras Activas e Inativas",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Gestion Carreras"
                ],
                "summary": "Ingrese los datos para crear una carrera",
                "description": "Ingrese los datos requeridos",
                "operationId": "Store Carreras",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "encargado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "encargado": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Creation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully Creation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/carreras/adminE": {
            "get": {
                "tags": [
                    "Gestion Carreras"
                ],
                "summary": "Lista de carreras activas",
                "description": "Carreras a Visualizar",
                "operationId": "Carreras Activas",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/carreras/admin/{id}": {
            "get": {
                "tags": [
                    "Gestion Carreras"
                ],
                "summary": " Información de una carrera según su id",
                "description": "Retorna la información de una carrera",
                "operationId": "Ver carrer por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese el id de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Gestion Carreras"
                ],
                "summary": "Actualizar la información de una carrera",
                "description": " Actualizar la información de una carrera",
                "operationId": " Actualizar Carreras",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese el id de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "encargado": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/carreras/desactiva/admin/{carreras}": {
            "get": {
                "tags": [
                    "Gestion Carreras"
                ],
                "summary": "Activar e Inactivar Carreras",
                "operationId": "2910c2c0718190ba283324ef0e036959",
                "parameters": [
                    {
                        "name": "carreras",
                        "in": "path",
                        "description": "Ingrese el id de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "name": "Authorization",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://backendedupoli.test/oauth/authorize",
                        "tokenUrl": "http://backendedupoli.test/oauth/token",
                        "refreshUrl": "http://backendedupoli.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}